New version
New version, with improved graphics, possibility of creating up to 100 X 100 cells with 9000 mines. Minimap for orientation of the game position on the board, new drop-down menu, equal level layout on both windows and mobile devices. Possibility of replaying the game on the same level without opening it as a new game. New design for the mines and seconds counters, allowing for a greater number of mines and seconds. Google Playstore is on the way. Removed dependencies from the sevendigitssgegment component

Translated with www.DeepL.com/Translator (free version)